+++ CHANGES.md
@@ -3,7 +3,7 @@
 ### Fixed
 
 - Pass the $CUDA_PATH/include path to the nvrtc compiler; otherwise it will not `#include` anything.
-- Work around `Ctypes.bigarray_start` because it does not support half precision.
+- Work around `Ctypes.bigarray_start` and `typ_of_bigarray_kind` because `ctypes` does not support half precision.
 
 ## [0.4.0] 2024-07-21